Move small classes into GeantSetup
